+++ src/screens/WalkThrough/constants.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: BUSL-1.1
+
+import {rem, screenWidth} from 'rn-units/index';
+
+export const CIRCLE_DIAMETER = screenWidth * 1.1;
+export const CIRCLE_PADDING_VERTICAL = rem(100);
+export const ANIMATION_CONFIG = {duration: 500};
+export const ANIMATION_DELAY = 300;
