+++ src/screens/WalkThrough/types.ts
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: BUSL-1.1
+
+export type DescriptionRenderData = {
+  type: 'text' | 'ice' | 'url';
+  value?: string;
+};
